+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.0.0] - 11-05-2025
+
+### Added
+- IntProperty, BoolProperty, and FloatProperty (src/property.hpp)
+- PropertyStore for AVR and ESP32 architectures (src/propertystore.hpp)
+
 ## [0.0.1] - 10-05-2025
 ### Added
 - This changelog